requirements.txt
pytest==7.1.2
requests==2.28.1
python-dotenv==0.21.0
voluptuous==0.13.1
pytest-voluptuous==1.2.0
curlify==2.2.1
allure-pytest==2.10.0